+++ GRDB/QueryInterface/SQLTableBuilder.swift
@@ -429,9 +429,9 @@ public final class ColumnDefinition {
     private var notNullConflictResolution: SQLConflictResolution?
     private var uniqueConflictResolution: SQLConflictResolution?
     private var checkConstraints: [_SQLExpression] = []
+    private var foreignKeyConstraints: [(table: String, column: String?, deleteAction: SQLForeignKeyAction?, updateAction: SQLForeignKeyAction?, deferred: Bool)] = []
     private var defaultExpression: _SQLExpression?
     private var collationName: String?
-    private var reference: (table: String, column: String?, deleteAction: SQLForeignKeyAction?, updateAction: SQLForeignKeyAction?, deferred: Bool)?
     
     init(name: String, type: SQLColumnType) {
         self.name = name
@@ -598,7 +598,7 @@ public final class ColumnDefinition {
     ///       See https://www.sqlite.org/foreignkeys.html#fk_deferred.
     /// - returns: Self so that you can further refine the column definition.
     public func references(table: String, column: String? = nil, onDelete deleteAction: SQLForeignKeyAction? = nil, onUpdate updateAction: SQLForeignKeyAction? = nil, deferred: Bool = false) -> ColumnDefinition {
-        reference = (table: table, column: column, deleteAction: deleteAction, updateAction: updateAction, deferred: deferred)
+        foreignKeyConstraints.append((table: table, column: column, deleteAction: deleteAction, updateAction: updateAction, deferred: deferred))
         return self
     }
     
@@ -655,7 +655,7 @@ public final class ColumnDefinition {
             chunks.append(collationName)
         }
         
-        if let (table, column, deleteAction, updateAction, deferred) = reference {
+        for (table, column, deleteAction, updateAction, deferred) in foreignKeyConstraints {
             chunks.append("REFERENCES")
             if let column = column {
                 chunks.append("\(table.quotedDatabaseIdentifier)(\(column.quotedDatabaseIdentifier))")
@@ -704,7 +704,7 @@ private struct IndexDefinition {
         chunks.append("\(table.quotedDatabaseIdentifier)(\((columns.map { $0.quotedDatabaseIdentifier } as [String]).joinWithSeparator(", ")))")
         if let condition = condition {
             chunks.append("WHERE")
-            var arguments: StatementArguments? = nil // nil so that checkExpression.sql(&arguments) embeds literals
+            var arguments: StatementArguments? = nil // nil so that condition.sql(&arguments) embeds literals
             chunks.append(condition.sql(&arguments))
         }
         return chunks.joinWithSeparator(" ")